requirements.txt
urllib3
rich
colorama
requests
bs4
jinja2
dnspython
aiohttp
whois
pyOpenSSL
python-nmap
aioquic
blessings